functions/trigger_firestore_cart_updated/requirements.txt
google-events
jsondiff